rename crc header and add portable rotl
